+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.0.0 - 2021-01-17
+
+### Added
+- Categories System
+- Comment System
+- Link Categories
+
+### Fixed
+- Title error [#2](https://github.com/greencheap/extension-blog/issues/2)
+
 ## 1.0.8 - 2020-09-14
 
 ### Remove